+++ src/game/MiscHandler.cpp
@@ -153,7 +153,7 @@ void WorldSession::HandleWhoOpcode( WorldPacket & recv_data )
                 continue;
 
             // player can see MODERATOR, GAME MASTER, ADMINISTRATOR only if CONFIG_GM_IN_WHO_LIST
-            if ((itr->second->GetSession()->GetSecurity() > gmLevelInWhoList))
+            if ((uint)itr->second->GetSession()->GetSecurity() > gmLevelInWhoList)
                 continue;
         }
